requirements.txt
flask==3.0.2
pymilvus==2.3.6
sentence-transformers==2.5.1
google-generativeai==0.3.2
python-dotenv==1.0.1 